jsc/core: port 32-bit Rust internals to i586 (pointer-width, tags, layout)
Make the Rust runtime correct on 32-bit pointers (i586 MinGW, no SSE) by
fixing pointer-width assumptions and NaN-boxed tag handling:

- JSValue / DecodedJSValue / CallFrame / FFI: 32-bit JSVALUE32_64 layout,
  u64 tag constants, pointer-sized field handling.
- windows_sys: WSADATA size, TEB/PEB segment registers for x86, extern
  signatures; semantic-version inline buffer always [u8; 8].
- node_os / node_process / socket / webcore / codegen libs: pointer-size,
  Architecture::X86, FFI sizes, 32-bit fixes.
- crash_handler CPU feature detection for x86; alloc BabyVec size assertion.

This is the 32-bit portability layer underpinning the Win9x/XP target.

+++ src/js_parser_jsc/Macro.rs
@@ -321,7 +321,7 @@ pub(crate) fn __bun_macro_context_call(
     // SAFETY: `data` is the `Box<MacroContext>` allocated in `init` above; the
     // lower-tier handle is uniquely borrowed for this call so no alias exists.
     let inner = unsafe { &mut *ctx.data.cast::<MacroContext>() };
-    inner.javascript_object = JSValue::from_encoded(ctx.javascript_object.0 as usize);
+    inner.javascript_object = JSValue::from_encoded(ctx.javascript_object.0 as u64);
     let caller_loc = caller.loc;
     inner
         .call(
